+++ src/analysisd/decoders/decoder.c
@@ -70,6 +70,7 @@ void DecodeEvent(Eventinfo *lf)
 #endif
 
         lf->decoder_info = nnode;
+        lf->log_after_prematch = pmatch;
         child_node = node->child;
 
         /* If no child node is set, set the child node
@@ -104,6 +105,8 @@ void DecodeEvent(Eventinfo *lf)
                         }
 
                         lf->decoder_info = nnode;
+                        lf->log_after_parent = pmatch;
+                        lf->log_after_prematch = cmatch;
 
                         break;
                     }