+++ src/config/localfile-config.c
@@ -11,6 +11,7 @@
 #include "localfile-config.h"
 #include "config.h"
 
+static void Free_Logreader(logreader * logf);
 
 int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
 {
@@ -57,7 +58,6 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
     }
 
     memset(logf + pl, 0, sizeof(logreader));
-    os_calloc(1, sizeof(wlabel_t), logf[pl].labels);
     logf[pl].ign = 360;
     os_calloc(2, sizeof(logsocket *), logf[pl].target_socket);
 
@@ -184,19 +184,23 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
                     glob_set = pl + 1;
                 }
 
-                if (glob(node[i]->content, 0, NULL, &g) != 0) {
+                switch (glob(node[i]->content, 0, NULL, &g)) {
+                case 0:
+                    break;
+                case GLOB_NOMATCH:
+                    mwarn(GLOB_NFOUND, node[i]->content);
+                    goto clean;
+                default:
                     merror(GLOB_ERROR, node[i]->content);
-                    os_strdup(node[i]->content, logf[pl].file);
-                    i++;
-                    continue;
+                    goto clean;
                 }
 
                 /* Check for the last entry */
                 if ((g.gl_pathv[glob_offset]) == NULL) {
                     /* Check when nothing is found */
                     if (glob_offset == 0) {
-                        merror(GLOB_NFOUND, node[i]->content);
-                        return (OS_INVALID);
+                        mwarn(GLOB_NFOUND, node[i]->content);
+                        goto clean;
                     }
                     i++;
                     continue;
@@ -233,24 +237,7 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
                     pl++;
                     os_realloc(logf, (pl +2)*sizeof(logreader), log_config->config);
                     logf = log_config->config;
-
-                    logf[pl].file = NULL;
-                    logf[pl].alias = NULL;
-                    logf[pl].logformat = NULL;
-                    logf[pl].fp = NULL;
-                    logf[pl].ffile = NULL;
-		            logf[pl].djb_program_name = NULL;
-		            logf[pl].labels = NULL;
-		            logf[pl].query = NULL;
-
-                    logf[pl +1].file = NULL;
-                    logf[pl +1].alias = NULL;
-                    logf[pl +1].logformat = NULL;
-		            logf[pl +1].fp = NULL;
-		            logf[pl +1].ffile = NULL;
-                    logf[pl +1].djb_program_name = NULL;
-                    logf[pl +1].labels = NULL;
-		            logf[pl +1].query = NULL;
+                    memset(logf + pl, 0, 2 * sizeof(logreader));
                 }
 
 
@@ -352,24 +339,77 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
         i++;
     }
 
+    if (logf[pl].target == NULL) {
+        os_calloc(2, sizeof(char*), logf[pl].target);
+        os_strdup("agent", logf[pl].target[0]);
+    }
+
     /* Validate glob entries */
     if (glob_set) {
         char *format;
+        char ** target;
+        char * outformat;
+        wlabel_t * labels;
 
         /* Get log format */
         if (logf[pl].logformat) {
-           format = logf[0].logformat;
+            format = logf[pl].logformat;
+            logf[pl].logformat = NULL;
         } else if (logf[glob_set - 1].logformat) {
-            format = logf[0].logformat;
+            format = logf[glob_set - 1].logformat;
+            logf[glob_set - 1].logformat = NULL;
         } else {
             merror(MISS_LOG_FORMAT);
+            return (OS_INVALID);
+        }
+
+        /* Get target */
+        if (logf[pl].target) {
+            target = logf[pl].target;
+            logf[pl].target = NULL;
+        } else if (logf[glob_set - 1].target) {
+            target = logf[glob_set - 1].target;
+            logf[glob_set - 1].target = NULL;
+        } else {
+            return (OS_INVALID);
+        }
+
+        // Delete target socket
+
+        if (logf[pl].target_socket) {
+            free(logf[pl].target_socket);
+        } else if (logf[glob_set - 1].target_socket) {
+            free(logf[glob_set - 1].target_socket);
+        } else {
+            return (OS_INVALID);
+        }
+
+        /* Get output format */
+        if (logf[pl].outformat) {
+            outformat = logf[pl].outformat;
+            logf[pl].outformat = NULL;
+        } else if (logf[glob_set - 1].outformat) {
+            outformat = logf[glob_set - 1].outformat;
+            logf[glob_set - 1].outformat = NULL;
+        } else {
+            outformat = NULL;
+        }
+
+        /* Get labels */
+        if (logf[pl].labels) {
+            labels = logf[pl].labels;
+            logf[pl].labels = NULL;
+        } else if (logf[glob_set - 1].labels) {
+            labels = logf[glob_set - 1].labels;
+            logf[glob_set - 1].labels = NULL;
+        } else {
             return (OS_INVALID);
         }
 
         /* The last entry is always null on glob */
         pl--;
 
-        /* Set format for all entries */
+        /* Set format and target for all entries */
         for (i = (glob_set - 1); i <= pl; i++) {
             /* Every entry must be valid */
             if (!logf[i].file) {
@@ -378,9 +418,44 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
             }
 
             if (logf[i].logformat == NULL) {
-		        os_strdup(format,logf[i].logformat);
+                os_strdup(format, logf[i].logformat);
+            }
+
+            if (logf[i].target == NULL) {
+                int j;
+
+                for (j = 0; target[j]; j++) {
+                    os_realloc(logf[i].target, sizeof(char *) * (j + 2), logf[i].target);
+                    os_strdup(target[j], logf[i].target[j]);
+                }
+
+                logf[i].target[j] = NULL;
+                os_calloc(j + 1, sizeof(logsocket *), logf[i].target_socket);
+            }
+
+            if (outformat && !logf[i].outformat) {
+                os_strdup(outformat, logf[i].outformat);
+            }
+
+            if (logf[i].labels == NULL) {
+                logf[i].labels = labels_dup(labels);
             }
         }
+
+        // Clear memory
+
+        for (i = 0; target[i]; i++) {
+            free(target[i]);
+        }
+
+        free(target);
+        free(format);
+        free(outformat);
+        labels_free(labels);
+    }
+
+    if (!logf[pl].labels) {
+        os_calloc(1, sizeof(wlabel_t), logf[pl].labels);
     }
 
     /* Missing log format */
@@ -393,11 +468,6 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
     if (!logf[pl].file) {
         merror(MISS_FILE);
         return (OS_INVALID);
-    }
-
-    if (logf[pl].target == NULL) {
-        os_calloc(2, sizeof(char*), logf[pl].target);
-        os_strdup("agent", logf[pl].target[0]);
     }
 
     /* Verify a valid event log config */
@@ -420,6 +490,11 @@ int Read_Localfile(XML_NODE node, void *d1, __attribute__((unused)) void *d2)
     }
 
     return (0);
+
+clean:
+    Free_Logreader(logf + pl);
+    memset(logf + pl, 0, sizeof(logreader));
+    return 0;
 }
 
 int Test_Localfile(const char * path){
@@ -441,26 +516,12 @@ int Test_Localfile(const char * path){
 }
 
 void Free_Localfile(logreader_config * config){
-    int i, j;
+    int i;
 
     if (config) {
         if (config->config) {
             for (i = 0; config->config[i].file; i++) {
-                free(config->config[i].ffile);
-                free(config->config[i].file);
-		        free(config->config[i].logformat);
-                free(config->config[i].djb_program_name);
-                free(config->config[i].alias);
-                free(config->config[i].query);
-                for (j = 0; config->config[i].target[j]; j++) {
-                    free(config->config[i].target[j]);
-                }
-                free(config->config[i].target);
-                labels_free(config->config[i].labels);
-                if (config->config[i].fp) {
-                    fclose(config->config[i].fp);
-                }
-                free(config->config[i].target_socket);
+                Free_Logreader(&config->config[i]);
             }
 
             free(config->config);
@@ -475,5 +536,35 @@ void Free_Localfile(logreader_config * config){
 
             free(config->socket_list);
         }
+    }
+}
+
+void Free_Logreader(logreader * logf) {
+    int i;
+
+    if (logf) {
+        free(logf->ffile);
+        free(logf->file);
+        free(logf->logformat);
+        free(logf->djb_program_name);
+        free(logf->alias);
+        free(logf->query);
+
+        if (logf->target) {
+            for (i = 0; logf->target[i]; i++) {
+                free(logf->target[i]);
+            }
+
+            free(logf->target);
+        }
+
+        labels_free(logf->labels);
+
+        if (logf->fp) {
+            fclose(logf->fp);
+        }
+
+        free(logf->target_socket);
+        free(logf->outformat);
     }
 }